project/requirements.txt
asyncio
selenium
bs4
fastapi
pandas
uvicorn[standard]
openpyxl
mysql-connector-python
json
django-mathfilters
django-address